Promote band-level coords to data var attributes for stackstac (#37)
Closes #36 

As mentioned in the issue since xpystac always returns an
xarray.Dataset, it needs to be careful about the conversion of coords
that have "band" as their dim. In a DataArray each version that
`stackstac` natively returns you end up with scalar coords per-band. I
think the best way to map that onto the data variables is by using
converting scalar coords to attrs.